code/requirements.txt
argcomplete==1.8.2
bleach==1.5.0
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
docx2txt==0.6
EbookLib==0.15
google-api-python-client==1.7.10
google-auth==1.6.3
google-auth-httplib2==0.0.3
gspread==3.1.0
httplib2==0.13.1
idna==2.8
lxml==4.4.0
numpy==1.17.0
oauth2client==4.1.3
pandas==0.25.0
pdfminer.six==20181108
Pillow==6.1.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycryptodome==3.8.2
PyPDF2==1.26.0
python-dateutil==2.8.0
python-pptx==0.6.5
pytz==2019.1
requests==2.22.0
rsa==4.0
six==1.10.0
sortedcontainers==2.1.0
uritemplate==3.0.0
urllib3==1.25.3
webencodings==0.5.1
XlsxWriter==1.1.8
